backend/requirements.txt
Django==3.2.16
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
gunicorn==20.1.0
Pillow==10.0.0
pytest==6.2.4
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.0
psycopg2-binary==2.9.3
PyYAML==6.0
flake8==6.0.0
webcolors==1.11.1